REF: Rename commit message checker module
